Update practice.js
Resolved following issue : When user is going to see the neighboring shlokas,and flips the card back it should show the original question shloka. eg. I open the flip page. It shows 1.4. I see neighboring shloka say 1.5, 1.6  and then if i swipe up then it shows adhyay 1 shloka 6 but question was adhyay 1 shloka 4. similarly if i move to next shloka and come back to this shloka, it will show adhya 1 shloka 6 instead of adhyay 1 shloka 4 which was the question.